refactor(money): extract shared shimmer primitive to remove duplication
Extract the shimmer sweep hook and gradient band into a shared
app/components/UI/Shimmer module and consume it from both TextShimmer
and the Card ShimmerOverlay, eliminating the duplicated animation core
flagged by SonarQube.
