+++ app/components/UI/Shimmer/index.ts
@@ -1,0 +1,4 @@
+export { default as ShimmerBand } from './ShimmerBand';
+export type { ShimmerBandProps } from './ShimmerBand';
+export { useShimmerSweep } from './useShimmerSweep';
+export type { ShimmerSweepOptions } from './useShimmerSweep';
